Editor goodness

+++ engine/src/Importer.cpp
@@ -556,7 +556,7 @@ struct ShaderHeader
     char     assetID            = (char)EFileType::SHADER;
     int      vertexPathLenght   = 0;
     int      fragmentPathLenght = 0;
-    uint16_t featureMask        = 0;
+    uint8_t  featureMask        = 0;
 };
 
 void GPE::writeShaderFile(const char* dst, const ShaderCreateConfig& arg)
